+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="br.ufmg.coltec.topicos_e04_backgroundtasks">
-
+    package="com.example.e04_topicos">
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <application
         android:allowBackup="true"
-        android:dataExtractionRules="@xml/data_extraction_rules"
-        android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.TopicosE04BackgroundTasks"
-        tools:targetApi="31">
+        android:theme="@style/Theme.E04_Topicos">
         <activity
             android:name=".MainActivity"
             android:exported="true">
