+++ app/src/main/java/com/example/e04_topicos/ImageDownloader.java
@@ -1,4 +1,4 @@
-package br.ufmg.coltec.topicos_e04_backgroundtasks;
+package com.example.e04_topicos;
 
 /**
  * Classe retirada do livro Google Android, 3a edição
