+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Topicos-E04-BackgroundTasks</string>
+    <string name="app_name">E04_Topicos</string>
 </resources>
