+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Topicos-E04-BackgroundTasks"
+rootProject.name = "E04_Topicos"
 include ':app'
